Refactor workflow implementation
